seperated more of the homemain file
