+++ src/pages/admin/datatable.js
@@ -1,4 +1,4 @@
-import Content from '../../components/homecontent';
+import Content from '../../components/homemain';
 
 export default function DataTables() {
   return <Content title="Data Tables" />;
