+++ src/pages/admin/inbox.js
@@ -1,4 +1,4 @@
-import Content from '../../components/homecontent';
+import Content from '../../components/homemain';
 
 export default function InboxPage() {
   return <Content title="Inbox" />;
