+++ src/pages/admin/settings.js
@@ -1,4 +1,4 @@
-import Content from '../../components/homecontent';
+import Content from '../../components/homemain';
 
 export default function SettingsPage() {
   return <Content title="Settings" />;
