+++ src/pages/admin/status.js
@@ -1,4 +1,4 @@
-import Content from '../../components/homecontent';
+import Content from '../../components/homemain';
 
 export default function StatusPage() {
   return <Content title="Data Monitoring &amp; Updates" />;
