+++ src/pages/admin/userpage.js
@@ -1,4 +1,4 @@
-import Content from '../../components/homecontent';
+import Content from '../../components/homemain';
 
 export default function UserPage() {
   return <Content title="User Page" />;
