+++ src/pages/index.js
@@ -1,4 +1,4 @@
-import Content from '../components/homecontent';
+import Content from '../components/homemain';
 
 export default function HomePage() {
   return <Content title="M-Dash Home" />;
